+++ include/cute.h
@@ -13,6 +13,7 @@ extern "C" {
 #include "collection/bitmap.h"
 #include "collection/bitset.h"
 #include "collection/vector.h"
+#include "collection/hashmap.h"
 
 #include "hash/hash.h"
 #include "macro.h"